+++ frontend/packages/manager/src/styles/common/ListStyles.jsx
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+const RowActions = styled.div`
+    display: flex;
+    flex-wrap: nowrap;
+    align-items: center;
+`;
+
+export { RowActions };